+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add warning to top of storefront when user is in development mode telling
+  changes to content can't be promoted
 
 ## [4.0.0] - 2019-06-18
 ### Added